# Conflicts:
#	data_juicer/ops/deduplicator/__init__.py
#	data_juicer/ops/filter/__init__.py
#	data_juicer/ops/mapper/__init__.py
#	data_juicer/utils/model_utils.py
#	environments/dev_requires.txt
#	environments/minimal_requires.txt
#	environments/science_requires.txt

+++ tests/ops/deduplicator/test_video_deduplicator.py
@@ -1,7 +1,7 @@
 import os
 import unittest
 
-from datasets import Dataset
+from data_juicer.core.data import NestedDataset as Dataset
 
 from data_juicer.ops.deduplicator.video_deduplicator import VideoDeduplicator
 from data_juicer.utils.unittest_utils import DataJuicerTestCaseBase
@@ -16,22 +16,29 @@ class VideoDeduplicatorTest(DataJuicerTestCaseBase):
     video3_path = os.path.join(data_path, 'video3.mp4')
     # video1_dup.mp4 is a duplicate sample of video1.mp4
     video4_path = os.path.join(data_path, 'video1_dup.mp4')
-    os.symlink(video1_path, video4_path)
+    if not os.path.exists(video4_path):
+        os.symlink(video1_path, video4_path)
     # video2_dup.mp4 is a duplicate sample of video2.mp4
     video5_path = os.path.join(data_path, 'video2_dup.mp4')
-    os.symlink(video2_path, video5_path)
+    if not os.path.exists(video5_path):
+        os.symlink(video2_path, video5_path)
     # video3_dup.mp4 is a duplicate sample of video3.mp4
     video6_path = os.path.join(data_path, 'video3_dup.mp4')
-    os.symlink(video3_path, video6_path)
+    if not os.path.exists(video6_path):
+        os.symlink(video3_path, video6_path)
     # video3_dup_dup.mp4 is a duplicate sample of video6.mp4
     video7_path = os.path.join(data_path, 'video3_dup_dup.mp4')
-    os.symlink(video6_path, video7_path)
+    if not os.path.exists(video7_path):
+        os.symlink(video6_path, video7_path)
 
     def _run_video_deduplicator(self, dataset: Dataset, target_list, op):
+        expected_keys = [op.video_key, op.text_key]
+        key_list = [key for key in expected_keys
+                    if len(target_list) > 0 and key in target_list[0]]
 
         dataset = dataset.map(op.compute_hash)
         dataset, _ = op.process(dataset)
-        dataset = dataset.select_columns(column_names=[op.video_key])
+        dataset = dataset.select_columns(column_names=key_list)
         res_list = dataset.to_list()
         self.assertEqual(res_list, target_list)
 
@@ -101,6 +108,50 @@ class VideoDeduplicatorTest(DataJuicerTestCaseBase):
         op = VideoDeduplicator()
         self._run_video_deduplicator(dataset, tgt_list, op)
 
+    def test_3_consider_text(self):
+
+        ds_list = [{
+            'videos': [self.video1_path],
+            'text': '<video> text1'
+        }, {
+            'videos': [self.video2_path],
+            'text': '<video> text2'
+        }, {
+            'videos': [self.video3_path],
+            'text': '<video> text3'
+        }, {
+            'videos': [self.video4_path],
+            'text': '<video> text1'
+        }, {
+            'videos': [self.video5_path],
+            'text': '<video> text5'
+        }, {
+            'videos': [self.video6_path],
+            'text': '<video> text3'
+        }, {
+            'videos': [self.video7_path],
+            'text': '<video> text7'
+        }]
+        tgt_list = [{
+            'videos': [self.video1_path],
+            'text': '<video> text1'
+        }, {
+            'videos': [self.video2_path],
+            'text': '<video> text2'
+        }, {
+            'videos': [self.video3_path],
+            'text': '<video> text3'
+        }, {
+            'videos': [self.video5_path],
+            'text': '<video> text5'
+        }, {
+            'videos': [self.video7_path],
+            'text': '<video> text7'
+        }]
+        dataset = Dataset.from_list(ds_list)
+        op = VideoDeduplicator(consider_text=True)
+        self._run_video_deduplicator(dataset, tgt_list, op)
+
     def test_4(self):
 
         ds_list = [{
@@ -119,6 +170,35 @@ class VideoDeduplicatorTest(DataJuicerTestCaseBase):
         }]
         dataset = Dataset.from_list(ds_list)
         op = VideoDeduplicator()
+        self._run_video_deduplicator(dataset, tgt_list, op)
+
+    def test_4_consider_text(self):
+
+        ds_list = [{
+            'videos': [self.video1_path, self.video2_path, self.video3_path],
+            'text': '<video> text1 <video> text2 <video> text3',
+        }, {
+            'videos': [self.video4_path, self.video5_path, self.video6_path],
+            'text': '<video> text1 <video> text2 <video> text3',
+        }, {
+            'videos': [self.video7_path, self.video5_path],
+            'text': '<video> text3 <video> text2',
+        }, {
+            'videos': [self.video6_path, self.video5_path],
+            'text': '<video> text6 <video> text2',
+        }]
+        tgt_list = [{
+            'videos': [self.video1_path, self.video2_path, self.video3_path],
+            'text': '<video> text1 <video> text2 <video> text3',
+        }, {
+            'videos': [self.video7_path, self.video5_path],
+            'text': '<video> text3 <video> text2',
+        }, {
+            'videos': [self.video6_path, self.video5_path],
+            'text': '<video> text6 <video> text2',
+        }]
+        dataset = Dataset.from_list(ds_list)
+        op = VideoDeduplicator(consider_text=True)
         self._run_video_deduplicator(dataset, tgt_list, op)
 
     def test_5(self):
@@ -143,6 +223,94 @@ class VideoDeduplicatorTest(DataJuicerTestCaseBase):
         }]
         dataset = Dataset.from_list(ds_list)
         op = VideoDeduplicator()
+        self._run_video_deduplicator(dataset, tgt_list, op)
+
+    def test_no_video(self):
+
+        ds_list = [{
+            'videos': [],
+            'text': '<video> text1'
+        }, {
+            'videos': [self.video2_path],
+            'text': '<video> text2'
+        }, {
+            'videos': [self.video3_path],
+            'text': '<video> text3'
+        }, {
+            'videos': [],
+            'text': '<video> text1'
+        }, {
+            'videos': [self.video5_path],
+            'text': '<video> text5'
+        }, {
+            'videos': [],
+            'text': '<video> text3'
+        }, {
+            'videos': [self.video7_path],
+            'text': '<video> text7'
+        }]
+        tgt_list = [{
+            'videos': [],
+            'text': '<video> text1'
+        }, {
+            'videos': [self.video2_path],
+            'text': '<video> text2'
+        }, {
+            'videos': [self.video3_path],
+            'text': '<video> text3'
+        }, {
+            'videos': [],
+            'text': '<video> text1'
+        }, {
+            'videos': [],
+            'text': '<video> text3'
+        }]
+        dataset = Dataset.from_list(ds_list)
+        op = VideoDeduplicator()
+        self._run_video_deduplicator(dataset, tgt_list, op)
+
+    def test_no_video_consider_text(self):
+
+        ds_list = [{
+            'videos': [],
+            'text': '<video> text1'
+        }, {
+            'videos': [self.video2_path],
+            'text': '<video> text2'
+        }, {
+            'videos': [self.video3_path],
+            'text': '<video> text3'
+        }, {
+            'videos': [],
+            'text': '<video> text1'
+        }, {
+            'videos': [self.video5_path],
+            'text': '<video> text5'
+        }, {
+            'videos': [],
+            'text': '<video> text3'
+        }, {
+            'videos': [self.video7_path],
+            'text': '<video> text3'
+        }]
+        tgt_list = [{
+            'videos': [],
+            'text': '<video> text1'
+        }, {
+            'videos': [self.video2_path],
+            'text': '<video> text2'
+        }, {
+            'videos': [self.video3_path],
+            'text': '<video> text3'
+        }, {
+            'videos': [self.video5_path],
+            'text': '<video> text5'
+        }, {
+            'videos': [],
+            'text': '<video> text3'
+        }]
+        dataset = Dataset.from_list(ds_list)
+        op = VideoDeduplicator(consider_text=True)
         self._run_video_deduplicator(dataset, tgt_list, op)
